allow nifti, openjpeg2 be system
