+++ console/CMakeLists.txt
@@ -1,12 +1,16 @@
 project(console)
 set(PROGRAMS dcm2niix)
 
+CONFIGURE_FILE(config.h.in config.h)
+include_directories(${CMAKE_CURRENT_BINARY_DIR})
+
 add_executable(dcm2niix
   main_console.cpp
   nii_dicom.cpp
   jpg_0XC3.cpp
   ujpeg.cpp
   nifti1_io_core.cpp
+  nifti1_io_ext.cpp
   nii_ortho.cpp
   nii_dicom_batch.cpp)
   
@@ -23,12 +27,22 @@ endif()
 
   
 #We now use miniz, removing zlib dependency
-#if(ZLIB_FOUND)
-#  TARGET_LINK_LIBRARIES(dcm2niix z)
-#else(ZLIB_FOUND)
-#  ADD_DEFINITIONS(-DmyDisableZlib)
-#endif(ZLIB_FOUND)
+if(ZLIB_FOUND)
+  ADD_DEFINITIONS(-DmyDisableMiniZ)
+  TARGET_LINK_LIBRARIES(dcm2niix ${ZLIB_LIBRARIES})
+endif(ZLIB_FOUND)
+if(NIFTI_FOUND)
+  include_directories(${NIFTI_INCLUDE_DIRS})
+  target_link_libraries(dcm2niix ${NIFTI_LIBRARIES})
+else()
+  include_directories(nifti)
+endif()
+if(OpenJPEG2_FOUND)
+  include_directories(${OpenJPEG2_INCLUDE_DIRS})
+  target_link_libraries(dcm2niix ${OpenJPEG2_LIBRARIES})
+else()
+  ADD_DEFINITIONS(-DmyDisableOpenJPEG)
+endif()
 ADD_DEFINITIONS(-DmyDisableJasper)
-ADD_DEFINITIONS(-DmyDisableOpenJPEG)
 
 install(TARGETS ${PROGRAMS} DESTINATION bin)
