+++ console/config.h.in
@@ -1,0 +1,6 @@
+#ifndef CONFIG_H
+#define CONFIG_H
+
+#cmakedefine NIFTI_FOUND
+
+#endif // CONFIG_H
