+++ console/nii_dicom_batch.cpp
@@ -35,11 +35,11 @@
 #ifdef myUseCOut
  #include <iostream>
 #endif
+#include "tinydir.h"
 #include "nifti1_io_core.h"
 #include "nifti1.h"
 #include "nii_dicom_batch.h"
 #include "nii_dicom.h"
-#include "tinydir.h"
 #include <ctype.h> //toupper
 #include <float.h>
 #include <math.h>
